vita/backend/requirements.txt
# Core
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.2
drf-spectacular>=0.27

# Auth & API
djangorestframework-simplejwt>=5.3
twilio>=9.0

# DB & cache
psycopg[binary]>=3.1
redis>=5.0
celery>=5.4
django-celery-beat>=2.6

# Storage (S3/MinIO)
boto3>=1.34
django-storages>=1.14

# Crypto & utils
cryptography>=42.0
python-dotenv>=1.0
dj-database-url>=2.2

# Dev / quality
pytest>=8.0
pytest-django>=4.8
ruff>=0.6
gunicorn>=22.0
